Move Tabs into the design system
