+++ web/src/features/evals/v2/components/Evaluators/Code/EvaluatorCodeLanguageSelector/EvaluatorCodeLanguageSelector.tsx
@@ -4,7 +4,7 @@ import {
   type EvalTemplateSourceCodeLanguage,
 } from "@langfuse/shared";
 
-import { Tabs } from "@/src/components/ui/tabs";
+import { Tabs } from "@/src/components/design-system/Tabs/Tabs";
 import { sourceCodeLanguageLabel } from "@/src/features/evals/v2/fns/evaluators/sourceCodeLanguageLabel";
 
 /** Selects the runtime language for a code evaluator. */
